Separete OVAL data structures
