+++ oscap_report/scap_results_parser/oval_definition_parser/info_of_test_parser.py
@@ -1,6 +1,6 @@
 import uuid
 
-from ..data_structures import OvalObject, OvalTest
+from ..data_structure_oval_node import OvalObject, OvalTest
 from ..namespaces import NAMESPACES
 
 MAX_MESSAGE_LEN = 99
